+++ src/lib/broadcast.ts
@@ -4,7 +4,7 @@ export function broadcast(
   frame: Window,
   payload: string,
   origin: string,
-  limitBroadCastToOrigin?: boolean
+  limitBroadcastToOrigin?: boolean
 ): void {
   let i = 0;
   let frameToBroadcastTo;
@@ -12,7 +12,7 @@ export function broadcast(
   try {
     frame.postMessage(payload, origin);
     if (hasOpener(frame) && frame.opener.top !== window.top) {
-      broadcast(frame.opener.top, payload, origin, limitBroadCastToOrigin);
+      broadcast(frame.opener.top, payload, origin, limitBroadcastToOrigin);
     }
 
     // previously, our max value was frame.frames.length
@@ -26,13 +26,13 @@ export function broadcast(
     while ((frameToBroadcastTo = frame.frames[i])) {
       // If a specifc `origin` is provided then we want to only broadcast messages
       // to domains that match the configured `origin`. Otherwise the browser will output noisy console errors.
-      if (origin !== "*" && limitBroadCastToOrigin) {
+      if (origin !== "*" && limitBroadcastToOrigin) {
         if (frameToBroadcastTo.origin !== origin) {
           i++;
           continue;
         }
       }
-      broadcast(frameToBroadcastTo, payload, origin, limitBroadCastToOrigin);
+      broadcast(frameToBroadcastTo, payload, origin, limitBroadcastToOrigin);
       i++;
     }
   } catch (_) {